libft.h: added the <stdint.h> library instead of defining it on libft.h

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: joesanto <joesanto@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/30 10:29:44 by joesanto          #+#    #+#             */
-/*   Updated: 2025/09/30 12:26:26 by joesanto         ###   ########.fr       */
+/*   Updated: 2025/09/30 12:49:54 by joesanto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 {
 	size_t	bytes;
 
-	if (size && nmemb > ULONG_MAX / size)
+	if (size && nmemb > SIZE_MAX / size)
 		return (0);
 	bytes = nmemb * size;
 	return (ft_memset(malloc(bytes), 0, bytes));
